+++ frontend/src/metabase/ui/components/icons/Icon/icons/index.ts
@@ -241,6 +241,8 @@ import mail_filled_component from "./mail_filled.svg?component";
 import mail_filled_source from "./mail_filled.svg?source";
 import metabot_component from "./metabot.svg?component";
 import metabot_source from "./metabot.svg?source";
+import metabot_sad_component from "./metabot_sad.svg?component";
+import metabot_sad_source from "./metabot_sad.svg?source";
 import metric_component from "./metric.svg?component";
 import metric_source from "./metric.svg?source";
 import model_component from "./model.svg?component";
@@ -888,6 +890,10 @@ export const Icons = {
   metabot: {
     component: metabot_component,
     source: metabot_source,
+  },
+  metabot_sad: {
+    component: metabot_sad_component,
+    source: metabot_sad_source,
   },
   metric: {
     component: metric_component,
